requirements.txt
flask==3.1.3
gunicorn==23.0.0
python-dotenv==1.2.1
matplotlib==3.10.8
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-tavily==0.2.17
langgraph==1.0.6
openai==2.15.0
httpx==0.28.1
